requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
sqlalchemy==2.0.15
psycopg2==2.9.3
python-dotenv==0.20.0
pydantic==1.10.5
